Renamed `helpers.ts` to `tyeguards.ts`
Also added `isCollection()`, `isSuccessCollection()` and `isFailureCollection()`
